# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/ioc": {
            "post": {
                "description": "Create New Ioc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "iocs"
                ],
                "summary": "Create New Ioc",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Ioc_input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ioc"
                        }
                    }
                }
            }
        },
        "/ioc/{id}": {
            "get": {
                "description": "Get 1 Ioc By id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "iocs"
                ],
                "summary": "Get 1 Ioc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id ioc",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ioc"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit Existing Ioc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "iocs"
                ],
                "summary": "Edit Ioc",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Ioc_input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ioc"
                        }
                    }
                }
            }
        },
        "/iocs": {
            "get": {
                "description": "get all data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "iocs"
                ],
                "summary": "Show an account",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ioc"
                        }
                    }
                }
            }
        },
        "/iocs/csv": {
            "get": {
                "description": "get data as csv",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "iocs"
                ],
                "summary": "Show data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ioc"
                        }
                    }
                }
            }
        },
        "/iocs/some": {
            "get": {
                "description": "get some data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "iocs"
                ],
                "summary": "Show some",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Ioc_Get"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ioc"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Ioc": {
            "type": "object",
            "required": [
                "ioc",
                "type"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ioc": {
                    "type": "string"
                },
                "otx_hash": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Ioc_Get": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.Ioc_input": {
            "type": "object",
            "required": [
                "ioc",
                "type"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "ioc": {
                    "type": "string"
                },
                "otx_hash": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    }
}